LayoutTests/imported/w3c/web-platform-tests/workers/data-url-shared-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'worker.port.onmessage = e => port.postMessage(e.data)')

Harness Error (TIMEOUT), message = null

FAIL application/javascript MIME allowed null is not an object (evaluating 'w.port.onmessage = t.step_func_done(function(e) {
      assert_equals(e.data, 'PASS');
    })')
FAIL text/plain MIME allowed null is not an object (evaluating 'w.port.onmessage = t.step_func_done(function(e) {
      assert_equals(e.data, 'PASS');
    })')
FAIL empty MIME allowed null is not an object (evaluating 'w.port.onmessage = t.step_func_done(function(e) {
      assert_equals(e.data, 'PASS');
    })')
FAIL communication goes both ways null is not an object (evaluating 'w.port.onmessage = t.step_func_done(function(e) {
      assert_equals(e.data, 'PASS');
    })')
FAIL indexedDB is present null is not an object (evaluating 'w.port.onmessage = t.step_func_done(function(e) {
      assert_equals(e.data, 'PASS');
    })')
FAIL indexedDB is inaccessible null is not an object (evaluating 'w.port.onmessage = t.step_func_done(function(e) {
      assert_equals(e.data, 'PASS');
    })')
FAIL cross-origin worker null is not an object (evaluating 'w.port.onmessage = t.step_func_done(function(e) {
      assert_equals(e.data, 'PASS');
    })')
FAIL worker has opaque origin null is not an object (evaluating 'w.port.onmessage = t.step_func_done(function(e) {
      assert_equals(e.data, 'PASS');
    })')
TIMEOUT A data: URL shared worker should not be shared among origins. Test timed out
